feat: update routing paths for login with role
