+++ example/components/Sidebar/SidebarContent.tsx
@@ -1,30 +1,38 @@
-import Link from 'next/link'
-import routes, { routeIsActive } from 'routes/sidebar'
-import * as Icons from 'icons'
-import { IIcon } from 'icons'
-import SidebarSubmenu from './SidebarSubmenu'
-import { useRouter } from 'next/router'
+import Link from "next/link";
+import routes, { routeIsActive, filterRoutesByRole } from "routes/sidebar";
+import * as Icons from "icons";
+import { IIcon } from "icons";
+import SidebarSubmenu from "./SidebarSubmenu";
+import { useRouter } from "next/router";
 
 function Icon({ icon, ...props }: IIcon) {
   // @ts-ignore
-  const Icon = Icons[icon]
-  return <Icon {...props} />
+  const Icon = Icons[icon];
+  return <Icon {...props} />;
 }
 
 interface ISidebarContent {
-  linkClicked: () => void
+  linkClicked: () => void;
 }
 
 function SidebarContent({ linkClicked }: ISidebarContent) {
-  const { pathname } = useRouter()
-  const appName = process.env.NEXT_PUBLIC_APP_NAME
+  const { pathname } = useRouter();
+  const appName = process.env.NEXT_PUBLIC_APP_NAME;
+
+  const role =
+    typeof window !== "undefined" ? localStorage.getItem("role") : null;
+
+  const filteredRoutes = filterRoutesByRole(routes, role || "");
 
   return (
     <div className="flex flex-col h-screen justify-between text-gray-500 dark:text-gray-400">
-      
       {/* Logo Atas (Tetap di Atas) */}
       <div className="flex items-center space-x-3 px-6 py-4 shrink-0">
-        <img src="/assets/img/logoatas.png" alt="Logo" className="w-auto h-auto" />
+        <img
+          src="/assets/img/logoatas.png"
+          alt="Logo"
+          className="w-auto h-auto"
+        />
         <span className="text-lg font-bold text-gray-800 dark:text-gray-200">
           {appName}
         </span>
@@ -33,7 +41,7 @@ function SidebarContent({ linkClicked }: ISidebarContent) {
       {/* Scrollable Menu */}
       <div className="flex-1 overflow-y-auto px-2">
         <ul>
-          {routes.map((route) =>
+          {filteredRoutes.map((route) =>
             route.routes ? (
               <SidebarSubmenu
                 route={route}
@@ -42,12 +50,12 @@ function SidebarContent({ linkClicked }: ISidebarContent) {
               />
             ) : (
               <li className="relative px-4 py-3" key={route.name}>
-                <Link href={route.path || '#'} scroll={false}>
+                <Link href={route.path || "#"} scroll={false}>
                   <a
                     className={`inline-flex items-center w-full text-sm font-semibold transition-colors duration-150 hover:text-gray-800 dark:hover:text-gray-200 ${
                       routeIsActive(pathname, route)
-                        ? 'dark:text-gray-100 text-gray-800'
-                        : ''
+                        ? "dark:text-gray-100 text-gray-800"
+                        : ""
                     }`}
                     onClick={linkClicked}
                   >
@@ -61,7 +69,7 @@ function SidebarContent({ linkClicked }: ISidebarContent) {
                     <Icon
                       className="w-5 h-5"
                       aria-hidden="true"
-                      icon={route.icon || ''}
+                      icon={route.icon || ""}
                     />
                     <span className="ml-4">{route.name}</span>
                   </a>
@@ -86,7 +94,7 @@ function SidebarContent({ linkClicked }: ISidebarContent) {
         </div>
       </div>
     </div>
-  )
+  );
 }
 
-export default SidebarContent
+export default SidebarContent;
