+++ pages/admin/perencanaan/delete.tsx
