+++ pages/admin/perencanaan/detail.tsx
