+++ pages/admin/pos/index.tsx
