+++ pages/admin/transaksi/detail.tsx
