+++ pages/superadmin/rekaptulasi/component/detail.tsx
